feat: Add multiple GPU support
Add support to manage multiple GPU's similar to CPU's

- Rename gpu to gpu memory across the board to make way for gpu min and
max values.
- Rename mem to memory to more descriptive.
- Make gpu memory a proper value in the host reports not and additional
attribute.
- Add setting and updating GPU and GPU memory counts via the API.
- GPU list is given to frames in RQD using the CUE_GPU_CORES env
variable.

Missing from the MR.

1) for simplicity I modified the Initial migration to incorporate all the changed needed in both the tables, functions and triggers.
To keep backward compatibility for users it will need to make it into a migration.

2) Our cuegui and rqd have diverged too much for easy merge. I've ported what I can, but it will likely be missing elements.

3) We don't use windows, the GPU RQD side uses nvidia-smi directly. we will want to find an OS-agnostic method.

4) tests. we will definitely need to write some tests.

+++ cuebot/src/main/java/com/imageworks/spcue/dao/oracle/ServiceDaoJdbc.java
@@ -60,7 +60,7 @@ public class ServiceDaoJdbc extends JdbcDaoSupport implements ServiceDao {
             s.minCores = rs.getInt("int_cores_min");
             s.maxCores = rs.getInt("int_cores_max");
             s.minMemory = rs.getLong("int_mem_min");
-            s.minGpu = rs.getLong("int_gpu_min");
+            s.minGpu = rs.getInt("int_gpu_min");
             s.threadable = rs.getBoolean("b_threadable");
             s.tags = splitTags(rs.getString("str_tags"));
             return s;
@@ -77,7 +77,7 @@ public class ServiceDaoJdbc extends JdbcDaoSupport implements ServiceDao {
             s.minCores = rs.getInt("int_cores_min");
             s.maxCores = rs.getInt("int_cores_max");
             s.minMemory = rs.getLong("int_mem_min");
-            s.minGpu = rs.getLong("int_gpu_min");
+            s.minGpu = rs.getInt("int_gpu_min");
             s.threadable = rs.getBoolean("b_threadable");
             s.tags = splitTags(rs.getString("str_tags"));
             s.showId = rs.getString("pk_show");
